Got rid of operator precedence and binding
Decided that operator precedence didn't really add anything of use.
Users must order reduction of rules by rule-precedence.
Decided that bind_left was clumsy. It's really something defined in
parsers that operate with a current token and a look-ahead token to
decide wether to shift or reduce. Earley doesn't work that way, so
instead a rule can now be declared to derive_right (or not), in which
case the parse tree will have it's weight to the right as opposed to the
default which is that the parse-tree has it's weight to the left. This
is explained in the manual.